eth transaction (#26)
* delete kotlin

* define API-based methods for blockchains

* delete method

* put method

* APi-based methods of ledger account

* fixed API url

* remove explicitly importing

* change class to LedgerAccount

* 20 seconds timeout

* change ledger account name

* fix bcash

* ledger customer

* order book

* add Record

* ledger Subscription

* ledger Transaction

* lombok compile

* API-bases methods for KMS

* lombok compile

* applied lombok

* add details for blockchain API, test code

* update ApiKey and BaseUrl.java

* update API details and test code

* update code

* update details for API

* update details for API

* add details to APIs and test code

* remove unused import package

* Ledger account test

* test create account for the ledger

* test create account for the ledger

* unit test ledger account

* unit test customer ledger

* unit test customer ledger

* unit test order book

* unit test

* unit test

* clean code

* validate objects

* validate objects

* validate CreateAccount

* validate CreateAccount

* validate object and unit test

* clean code

* transaction for bitcoin and litecoin

* apply async

* apply async

* change names

* sign KMS transaction

* bcach transaction

* fix litecoin

* fix errors

* xrp transaction

* xrp with kms transaction

* send xrp transaction

* xrp wallet

* btc wallet

* change test package

* bip32

* derivation path

* refactor code

* refactor code

* fix input parameter

* don not use singleton objects to prevent from risk of race condition

* clean code

* generate private keys and addresses

* ethe wallet

* ethe wallet

* bch, vet wallets

* eth private key

* eth transaction

+++ src/main/java/io/tatum/blockchain/Ethereum.java
@@ -9,6 +9,7 @@ import io.tatum.utils.BaseUrl;
 
 import java.io.IOException;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.concurrent.ExecutionException;
 
 public final class Ethereum {
@@ -24,9 +25,9 @@ public final class Ethereum {
     /**
      * For more details, see <a href="https://tatum.io/apidoc#operation/EthGetTransactionCount" target="_blank">Tatum API documentation</a>
      */
-    public BigDecimal ethGetTransactionsCount(String address) throws ExecutionException, InterruptedException {
+    public BigInteger ethGetTransactionsCount(String address) throws ExecutionException, InterruptedException {
         String uri = BaseUrl.getInstance().getUrl() + "/v3/ethereum/transaction/count/" + address;
-        return Async.get(uri, BigDecimal.class);
+        return Async.get(uri, BigInteger.class);
     }
 
     /**
